Implement Keychain::Item#password
For getting the password of an item; some of the behaviour
is not defined yet because I need to play with it more
first.

I also needed to switch to using the Foundation framework
so that the NSUTF8StringEncoding constant would be
defined.

Spect tests are also included for this new method.