Change argument style to reflect Ruby 1.9 convenience
I was unaware of the awesome new feature in Ruby 1.9 where the
last argument of a method would be assumed to be a Hash if there is an argument that looks like a Hash at the end of an argument
list.

The let me remove a bunch of braces, making the code a bit cleaner.

The next step is to replace the use of contstants with symbols to
use the other big Hash improvement in Ruby 1.9.